+++ app/src/main/java/com/example/app/util/FoodRatios.java
@@ -19,24 +19,24 @@ public class FoodRatios {
      * Creates and initializes the Feeding meals
      */
     public static void createFoodObjects() {
-        HashMap<String, Integer> breakfastIngredients = new HashMap<>();
-        breakfastIngredients.put("Flour", 0);
-        breakfastIngredients.put("Sugar", 0);
-        breakfastIngredients.put("Water", 0);
+        HashMap<String, Double> breakfastIngredients = new HashMap<>();
+        breakfastIngredients.put("Flour", 0.0);
+        breakfastIngredients.put("Sugar", 0.0);
+        breakfastIngredients.put("Water", 0.0);
         mUji  = new Food("Uji", breakfastIngredients);
 
-        HashMap<String, Integer> githeriIngredients = new HashMap<>();
-        githeriIngredients.put("Maize", 0);
-        githeriIngredients.put("Beans", 0);
-        githeriIngredients.put("Salt", 0);
-        githeriIngredients.put("Oil", 0);
+        HashMap<String, Double> githeriIngredients = new HashMap<>();
+        githeriIngredients.put("Maize", 0.0);
+        githeriIngredients.put("Beans", 0.0);
+        githeriIngredients.put("Salt", 0.0);
+        githeriIngredients.put("Oil", 0.0);
         mGitheri = new Food("Githeri", githeriIngredients);
 
-        HashMap<String, Integer> rice_beansIngredients = new HashMap<>();
-        rice_beansIngredients.put("Rice", 0);
-        rice_beansIngredients.put("Beans", 0);
-        rice_beansIngredients.put("Salt", 0);
-        rice_beansIngredients.put("Oil", 0);
+        HashMap<String, Double> rice_beansIngredients = new HashMap<>();
+        rice_beansIngredients.put("Rice", 0.0);
+        rice_beansIngredients.put("Beans", 0.0);
+        rice_beansIngredients.put("Salt", 0.0);
+        rice_beansIngredients.put("Oil", 0.0);
         mRiceBeans = new Food("Rice&Beans", rice_beansIngredients);
     }
 
@@ -49,9 +49,9 @@ public class FoodRatios {
 
     public static Food getUjiRatios(int attendance) {
         createFoodObjects();
-        HashMap<String, Integer> ingredients = mUji.getRatios();
-        ingredients.put("Flour", attendance / 15);
-        int flourQuantity = (int) ingredients.get("Flour");
+        HashMap<String, Double> ingredients = mUji.getRatios();
+        ingredients.put("Flour", attendance / 15.0);
+        double flourQuantity = (double) ingredients.get("Flour");
         ingredients.put("Sugar", flourQuantity / 4);
         ingredients.put("Water", flourQuantity * 2);
         return mUji;
@@ -65,12 +65,12 @@ public class FoodRatios {
      */
     public static Food getGitheriRatios(int attendance) {
         createFoodObjects();
-        HashMap<String, Integer> ingredients = mGitheri.getRatios();
-        int maizeQuantity = attendance / 9;
+        HashMap<String, Double> ingredients = mGitheri.getRatios();
+        double maizeQuantity = attendance / 9;
         ingredients.put("Maize", maizeQuantity);
         ingredients.put("Beans", maizeQuantity / 2);
-        ingredients.put("Salt", attendance / 400);
-        ingredients.put("Oil", attendance / 250);
+        ingredients.put("Salt", attendance / 400.0);
+        ingredients.put("Oil", attendance / 250.0);
         return mGitheri;
     }
 
@@ -82,18 +82,18 @@ public class FoodRatios {
      */
     public static Food getRiceBeansRatios(int attendance) {
         createFoodObjects();
-        HashMap<String, Integer> ingredients = mRiceBeans.getRatios();
-        int riceQuantity = attendance / 12;
+        HashMap<String, Double> ingredients = mRiceBeans.getRatios();
+        double riceQuantity = attendance / 12;
         ingredients.put("Rice", riceQuantity);
         ingredients.put("Beans", riceQuantity / 2);
-        ingredients.put("Salt", attendance / 400);
-        ingredients.put("Oil", attendance / 250);
+        ingredients.put("Salt", attendance / 400.0);
+        ingredients.put("Oil", attendance / 250.0);
         return mRiceBeans;
     }
 
     public static ArrayList<Pair> populateList(Food food) {
         ArrayList<Pair> mRatioPairs = new ArrayList<>();
-        HashMap<String, Integer> ratios = food.getRatios();
+        HashMap<String, Double> ratios = food.getRatios();
         Iterator iterator = ratios.entrySet().iterator();
         while(iterator.hasNext()) {
             Map.Entry pair = (Map.Entry) iterator.next();
